Added optional shard prefix
This allows ids to be generated with an optional prefix that can be used
for sharding. Both prefix length and value can be set from the config
file. While not perfect (that would allow us to dynamically alter an
individual id to a specific pool) it will at least allow us to route to
a specific cluster if need be.